% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce_join.R
\name{coalesce_join}
\alias{coalesce_join}
\title{Coalesce Join Two Data Frames}
\usage{
coalesce_join(
  x,
  y,
  by = NULL,
  suffix = c(".x", ".y"),
  join = dplyr::full_join,
  ...
)
}
\arguments{
\item{x}{The first data frame.}

\item{y}{The second data frame.}

\item{by}{A vector of variables to join by. If NULL, the default, joins by common key variables.}

\item{suffix}{A character vector of length 2 to add to the names of the columns in \code{x} and \code{y}.}

\item{join}{The join function to use, by default \code{dplyr::full_join}.}

\item{...}{Additional arguments to pass to the \code{join} function.}
}
\value{
A joined data frame with coalesced columns.
}
\description{
This function performs a full join between two data frames and then coalesces columns with the same name.
It prioritizes the left-hand side data frame in the case of NA values.
}
\examples{
df1 <- data.frame(a = 1:3, b = c(NA, NA, 3))
df2 <- data.frame(a = 2:4, b = 2:4)
coalesce_join(df1, df2)

}
