% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combiner.R
\name{combiner}
\alias{combiner}
\title{Combiner Function for Merging Duplicate Data in List of Data Frames}
\usage{
combiner(dfs)
}
\arguments{
\item{dfs}{A named list of data frames.}
}
\value{
A named list of data frames with duplicates combined and distinct rows.
}
\description{
This function detects duplicate data in a list of data frames and merges them.
}
