% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_by_closest_date.R
\name{join_list_of_dfs_to_timepoints}
\alias{join_list_of_dfs_to_timepoints}
\title{Join a list of dataframes to timepoints based on the closest date.}
\usage{
join_list_of_dfs_to_timepoints(
  pidns_and_timepoints,
  list_of_dfs,
  max_day_difference,
  remove_empty_option = TRUE
)
}
\arguments{
\item{pidns_and_timepoints}{A dataframe with PIDN (patient IDs) and DCDate (date/time points).}

\item{list_of_dfs}{List of dataframes that need to be joined with \code{pidns_and_timepoints}.}

\item{max_day_difference}{Maximum day difference allowed for the join on date.}

\item{remove_empty_option}{Logical, if TRUE, will remove empty rows and columns from the final tibble.}
}
\value{
A consolidated tibble after joining on the closest date.
}
\description{
This function takes a primary dataframe with patient IDs and timepoints, and joins it with a list of
dataframes on the closest date. After joining, duplicates are removed and the resulting dataframes are
consolidated into one tibble.
}
