% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_all_files.R
\name{read_all_files}
\alias{read_all_files}
\title{Read and Clean Multiple Files}
\usage{
read_all_files(
  filenames,
  prefix = NULL,
  skips = 0,
  clean_names = TRUE,
  qualtrics = FALSE,
  col_types = NULL
)
}
\arguments{
\item{filenames}{A character vector of file paths.}

\item{prefix}{An optional prefix for the output list names.}

\item{skips}{Number of rows to skip when reading the file.}

\item{clean_names}{Logical, whether to clean column names using \code{janitor::make_clean_names}.}

\item{qualtrics}{Logical, whether the CSV is in Qualtrics format.}

\item{col_types}{Column types, passed to \code{readxl::read_excel}.}
}
\value{
A list of cleaned data frames.
}
\description{
This function reads in multiple files (CSV and Excel), cleans the column names,
and combines the data into a list.
}
