% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_berlin.R
\name{score_berlin}
\alias{score_berlin}
\title{Berlin Sleep Scoring Function}
\usage{
score_berlin(data)
}
\arguments{
\item{data}{A data frame containing the Berlin questionnaire fields.}
}
\value{
A data frame with additional columns for scoring and risk level.
}
\description{
This function scores sleep quality based on the Berlin questionnaire.
}
\examples{
# Prepare some example data and score it
example_data <- data.frame(
  do_you_snore = c("yes", "no", "yes"),
  your_snoring_is = c("Louder than talking", NA, "normal"),
  how_often_do_you_snore = c("3-4 times a week", NA, "never"),
  bmi = c(32, 25, 28)
)
scored_data <- score_berlin(example_data)
}
